# api/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Admin:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      email:
        type: string
      first_name:
        type: string
      id:
        type: string
      is_active:
        type: boolean
      is_super:
        type: boolean
      last_name:
        type: string
      phone:
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@example.com
    name: MarketKit
  description: REST API for the MarketKit marketplace platform. Admin routes require
    a Bearer JWT token. User routes require a separate User Bearer JWT token.
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: MarketKit API
  version: "1.0"
paths:
...
  /admins:
    get:
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Items per page
        in: query
        name: limit
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminAuth: []
      summary: List all admin accounts
      tags:
      - Admins
  /admins/{id}:
    delete:
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminAuth: []
      summary: Permanently delete an admin (super-admin only)
      tags:
      - Admins
  /admins/{id}/super:
    patch:
      consumes:
      - application/json
      parameters:
      - description: Admin ID
        in: path
        name: id
        required: true
        type: string
      - description: '{\'
        in: body
        name: body
        required: true
        schema:
          additionalProperties:
            type: boolean
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Admin'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - AdminAuth: []
      summary: Grant or revoke super-admin status (super-admin only)
      tags:
      - Admins
...
securityDefinitions:
  AdminAuth:
    description: Admin JWT — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
  UserAuth:
    description: User JWT — prefix with "Bearer "
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"